requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchserve>=0.8.0
torch-model-archiver>=0.8.0
Pillow>=9.0.0
PyYAML>=6.0.0
requests>=2.28.0

